+++ spec/unit/imap/backup/account/folder_spec.rb
@@ -300,20 +300,33 @@ module Imap::Backup
         expect(client).to have_received(:select)
       end
 
-      it "set the flag" do
+      it "sets the flag" do
         expect(client).
-          to have_received(:uid_store).with([99], "+FLAGS", [:Foo])
+          to have_received(:uid_store).with([99], "FLAGS", [:Foo])
       end
     end
 
-    describe "#unset_flags" do
-      before { subject.unset_flags([99], [:Foo]) }
+    describe "#add_flags" do
+      before { subject.add_flags([99], [:Foo]) }
 
       it "uses select to have read-write access" do
         expect(client).to have_received(:select)
       end
 
-      it "set the flag" do
+      it "sets the flag" do
+        expect(client).
+          to have_received(:uid_store).with([99], "+FLAGS", [:Foo])
+      end
+    end
+
+    describe "#remove_flags" do
+      before { subject.remove_flags([99], [:Foo]) }
+
+      it "uses select to have read-write access" do
+        expect(client).to have_received(:select)
+      end
+
+      it "unsets the flag" do
         expect(client).
           to have_received(:uid_store).with([99], "-FLAGS", [:Foo])
       end