+++ lib/imap/backup/setup/account.rb
@@ -36,14 +36,15 @@ module Imap::Backup
         header menu
         modify_email menu
         modify_password menu
-        modify_backup_path menu
-        choose_folders menu
-        toggle_mirror_mode menu
-        modify_multi_fetch_size menu
         modify_server menu
         modify_connection_options menu
-        toggle_reset_seen_flags_after_fetch menu
         test_connection menu
+        toggle_mirror_mode menu
+        modify_backup_path menu
+        toggle_folder_blacklist menu
+        choose_folders menu
+        modify_multi_fetch_size menu
+        toggle_reset_seen_flags_after_fetch menu
         delete_account menu
         menu.choice("(q) return to main menu") { throw :done }
         menu.hidden("quit") { throw :done }
@@ -74,8 +75,17 @@ module Imap::Backup
       end
     end
 
+    def toggle_folder_blacklist(menu)
+      menu_item = "toggle folder inclusion mode (whitelist/backlist)"
+      new_value = account.folder_blacklist ? nil : true
+      menu.choice(menu_item) do
+        account.folder_blacklist = new_value
+      end
+    end
+
     def choose_folders(menu)
-      menu.choice("choose backup folders") do
+      action = account.folder_blacklist ? "exclude from backups" : "include in backups"
+      menu.choice("choose folders to #{action}") do
         Setup::FolderChooser.new(account).run
       end
     end