+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 * Removed 'Experimental' warning from 'migrate' command,
-* Removed 'Experimental' warning from 'export-to-thunderbird' command.
+* Removed 'Experimental' warning from 'export-to-thunderbird' command,
+* Renamed Folder#*_flags methods.
 
 ## [8.0.0.rc1] - 2022-09-19