+++ CHANGELOG.md
@@ -6,11 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## ??? - TBA
 
+### Added
+
+* An account option 'folder_blacklist'. When set the user chooses
+  which folders to **exclude** from backups. Otherwise, the folders
+  chosen by the user are the ones to **include**.
+
 ### Changed
 
 * Removed 'Experimental' warning from 'migrate' command,
 * Removed 'Experimental' warning from 'export-to-thunderbird' command,
-* Renamed Folder#*_flags methods.
+* Renamed Folder#*_flags methods,
+* Improved setup account menu ordering.
 
 ### Removed