+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 8.0.1 - 2022-09-24
+
+### Changed
+
+* Filtered out non-standard flags from `migrate` and `restore`.
+
 ## 8.0.0 - 2022-09-24
 
 ### Added